+++ retz-client/src/main/java/io/github/retz/cli/CommandRun.java
@@ -34,28 +34,23 @@ import java.util.Properties;
 
 public class CommandRun implements SubCommand {
     static final Logger LOG = LoggerFactory.getLogger(CommandRun.class);
-
-    @Parameter(names = "-cmd", required = true, description = "Remote command")
-    private String remoteCmd;
-
-    @Parameter(names = {"-A", "--appname"}, required = true, description = "Application name you loaded")
-    private String appName;
-
     @Parameter(names = {"-E", "--env"},
             description = "Pairs of environment variable names and values, like '-E ASAKUSA_M3BP_OPTS='-Xmx32g' -E SPARK_CMD=path/to/spark-cmd'")
     List<String> envs;
-
     @Parameter(names = "-cpu", description = "Number of CPU cores assigned to the job")
     int cpu = 1;
-
     @Parameter(names = "-mem", description = "Number of size of RAM(MB) assigned to the job")
     int mem = 32;
-
     @Parameter(names = "-gpu", description = "Number of GPU cards assigned to the job")
     int gpu = 0;
-
+    @Parameter(names = "-ports", description = "Number of ports (up to 1000) required to the job; Ports will be given as $PORT0, $PORT1, ...")
+    int ports = 0;
     @Parameter(names = "-stderr", description = "Print stderr after the task finished to standard error")
     boolean stderr = false;
+    @Parameter(names = "-cmd", required = true, description = "Remote command")
+    private String remoteCmd;
+    @Parameter(names = {"-A", "--appname"}, required = true, description = "Application name you loaded")
+    private String appName;
 
     @Override
     public String description() {
@@ -71,8 +66,11 @@ public class CommandRun implements SubCommand {
     public int handle(ClientCLIConfig fileConfig) {
         Properties envProps = SubCommand.parseKeyValuePairs(envs);
 
-        Job job = new Job(appName, remoteCmd,
-                envProps, cpu, mem, gpu);
+        if (ports < 0 || 1000 < ports) {
+            LOG.error("-ports must be within 0 to 1000: {} given.", ports);
+            return -1;
+        }
+        Job job = new Job(appName, remoteCmd, envProps, cpu, mem, gpu, ports);
 
         try (Client webClient = Client.newBuilder(fileConfig.getUri())
                 .enableAuthentication(fileConfig.authenticationEnabled())