+++ retz-common/src/main/java/io/github/retz/mesos/Resource.java
@@ -16,31 +16,31 @@
  */
 package io.github.retz.mesos;
 
+import io.github.retz.protocol.data.Range;
 import org.apache.mesos.Protos;
 
-import java.util.LinkedHashMap;
+import java.util.Arrays;
+import java.util.LinkedList;
 import java.util.List;
-import java.util.Map;
 import java.util.stream.Collectors;
 
 public class Resource {
 
     private double cpu;
     private int memMB;
-    private int diskMB;
-    private int reservedDiskMB;
     private int gpu;
-    private final Map<String, Protos.Resource> volumes = new LinkedHashMap<>();
+    private List<Range> ports;
+    private int diskMB;
 
     public Resource(double cpu, int memMB, int diskMB) {
-        this(cpu, memMB, diskMB, 0, 0);
+        this(cpu, memMB, diskMB, 0, new LinkedList<>());
     }
 
-    public Resource(double cpu, int memMB, int diskMB, int reservedDiskMB, int gpu) {
+    public Resource(double cpu, int memMB, int diskMB, int gpu, List<Range> ports) {
         this.cpu = cpu;
         this.memMB = memMB;
+        this.ports = ports;
         this.diskMB = diskMB;
-        this.reservedDiskMB = reservedDiskMB;
         this.gpu = gpu;
     }
 
@@ -48,9 +48,16 @@ public class Resource {
         this.cpu += rhs.cpu();
         this.memMB += rhs.memMB();
         this.diskMB += rhs.diskMB();
-        this.reservedDiskMB += rhs.reservedDiskMB();
+        for (Range lhs : ports) {
+            for (Range r : rhs.ports()) {
+                if (lhs.overlap(r)) {
+                    throw new RuntimeException("Port range overlapping: " + lhs + " and " + r);
+                }
+            }
+        }
+        this.ports.addAll(rhs.ports());
+        this.ports.sort( (l, r) -> r.getMin() - l.getMin());
         this.gpu += rhs.gpu();
-        this.volumes.putAll(rhs.volumes());
     }
 
     public double cpu() {
@@ -65,48 +72,98 @@ public class Resource {
         return diskMB;
     }
 
-    public int reservedDiskMB() {
-        return reservedDiskMB;
+    public List<Range> ports() {
+        return ports;
+    }
+
+    public int lastPort() {
+        if (ports.isEmpty()) {
+            return 0;
+        }
+        return ports.get(ports.size() - 1).getMax();
+    }
+
+    public int portAmount() {
+        return ports.stream().mapToInt(range -> range.getMax() - range.getMin() + 1).sum();
     }
 
     public int gpu() {
         return gpu;
     }
 
-    public Map<String, Protos.Resource> volumes() {
-        return volumes;
-    }
-
-    public String getPersistentVolumePath(String volumeId, String defaultPath) {
-        if (!volumes.containsKey(volumeId)) {
-            // volume not found
-            return defaultPath;
+    public Resource cut(int cpu, int memMB, int gpus, int ports, int lastPort) {
+        List<Range> ranges = new LinkedList<>();
+        int sum = 0;
+        for(Range range : this.ports) {
+            if (range.getMax() <= lastPort) {
+                continue;
+            } else if (sum >= ports) {
+                break;
+            }
+            int start = 0;
+            if (lastPort < range.getMin()) {
+                start = range.getMin();
+            } else {
+                start = lastPort + 1;
+            }
+            int end = 0;
+            if (range.getMax() - start > ports - sum) {
+                end = start + ports - sum - 1;
+            } else {
+                end = range.getMax();
+            }
+            sum += (end - start) + 1;
+            if (start <= end) {
+                ranges.add(new Range(start, end));
+            }
         }
-        Protos.Resource r = volumes.get(volumeId);
-        return r.getDisk().getVolume().getContainerPath();
+        return new Resource(cpu, memMB, 0, gpus, ranges);
     }
 
-    public void subCPU(double cpu) {
-        this.cpu -= cpu;
+    public List<Protos.Resource> construct() {
+        List<Protos.Resource> list = new LinkedList<Protos.Resource>();
+        list.add(Protos.Resource.newBuilder()
+                .setName("cpus")
+                .setScalar(Protos.Value.Scalar.newBuilder().setValue(cpu).build())
+                .setType(Protos.Value.Type.SCALAR)
+                .build());
+        list.add(Protos.Resource.newBuilder()
+                .setName("mem")
+                .setScalar(Protos.Value.Scalar.newBuilder().setValue(memMB))
+                .setType(Protos.Value.Type.SCALAR)
+                .build());
+        if (diskMB > 0) {
+            list.add(Protos.Resource.newBuilder()
+                    .setName("disk")
+                    .setScalar(Protos.Value.Scalar.newBuilder().setValue(diskMB))
+                    .setType(Protos.Value.Type.SCALAR)
+                    .build());
+        }
+        if (gpu > 0) {
+            list.add(Protos.Resource.newBuilder()
+                    .setName("gpus")
+                    .setScalar(Protos.Value.Scalar.newBuilder().setValue(gpu))
+                    .setType(Protos.Value.Type.SCALAR)
+                    .build());
+        }
+        if (! ports.isEmpty()) {
+            list.add(Protos.Resource.newBuilder()
+                    .setName("ports")
+                    .setRanges(Protos.Value.Ranges.newBuilder().addAllRange(
+                            ports.stream().map(range ->
+                                    Protos.Value.Range.newBuilder()
+                                            .setBegin(range.getMin())
+                                            .setEnd(range.getMax())
+                                            .build()
+                            ).collect(Collectors.toList())).build())
+                    .setType(Protos.Value.Type.RANGES)
+                    .build());
+        }
+        return list;
     }
-
-    public void subMemMB(int memMB) {
-        this.memMB -= memMB;
-    }
-
-    public void subGPU(int gpu) {
-        this.gpu -= gpu;
-    }
-
     @Override
     public String toString() {
-        List<String> vols = volumes.entrySet().stream()
-                .map(entry -> String.format("%s=%fMB:%s:%s",
-                        entry.getKey(), entry.getValue().getScalar().getValue(),
-                        entry.getValue().getDisk().getVolume().getContainerPath(),
-                        entry.getValue().getDisk().getVolume().getHostPath()))
-                .collect(Collectors.toList());
-        return String.format("cpu=%.1f, mem=%d, disk=%d, reserved_disk=%d, gpus=%d, volumes=[%s]",
-                cpu, memMB, diskMB, reservedDiskMB, gpu, String.join(", ", vols));
+        String portRanges = String.join(", ", ports.stream().map(port -> port.toString()).collect(Collectors.toList()));
+        return String.format("cpus=%.1f, mem=%dMB, disk=%dMB, gpu=%d, ports=[%s]", cpu, memMB, diskMB, gpu, portRanges);
     }
 }